requirements.txt
python-dotenv~=1.0.1
mistralai~=1.5.0
